Fix error for zero in-arity operators
Popping the stack for zero elements failed to return the right response
when the stack was non-empty. The negative index slices used to split
the stack did not work properly for n == 0, as there is no difference
between 0 and -0. This results in popping the full stack, which works if
empty, but not otherwise.